fix: bookmarks header menu
